+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### 2.11.10
 
 - Fix bug it should compare pbx versions correctly
+- Disable non-tls or plain http connection for lpc and avatar image html
 - Save getProductInfo webphone.useragent to local storage and use it in the next sip login
 - Save getProductInfo webphone.pn_expires to local storage and use it in the next incoming PN to check if the pn expired. If it is not present, use 50000 as default 50 seconds
 - Update call buttons via sip header X-WEBPHONE-CALL. The config will be merged one by one each time it receives